added draggable and onDragEnd for colored squares

+++ lib/cards.dart
@@ -8,10 +8,10 @@ class MatchCard {
   int redColor = 0;
   int greenColor = 0;
   int blueColor = 0;
-  int margin = 0;
+  double margin = 0;
 
 
-  MatchCard(int red, int green, int blue, int marginTop) {
+  MatchCard(int red, int green, int blue, double marginTop) {
     redColor = red;
     greenColor = green;
     blueColor = blue;
